on some systems isdigit() is a macro using an array. Cast its argument to int to keep gcc happy in this case.

+++ src/comLib/os/posix/xesClient.c
@@ -108,7 +108,7 @@ static int clientsock(host, port)
     server.sin_family = AF_INET;
     server.sin_port = htons(port);
 
-    if (isdigit(host[0]))
+    if (isdigit((int)host[0]))
 	server.sin_addr.s_addr = inet_addr(host);
     else {
